# testserver: set default rf to 1 for single node servers